fix: consolidate factorysession packages for pkg-file-count CI gate
Merge canonical event, result projection, and start source modules into
existing factorysessionexecution owners and fold checkpoint_store into
checkpoints so both packages stay within the 15-file limit.
