--- pkg/factorysessionexecution/canonical_events_test.go
@@ -1,125 +1,0 @@
-package factorysessionexecution
-
-import (
-	"encoding/json"
-	"errors"
-	"testing"
-	"time"
-)
-
-func TestBuildCanonicalSessionEvents_RunningAndTerminalSessions(t *testing.T) {
-	startedAt := time.Date(2026, 6, 8, 14, 0, 0, 0, time.UTC)
-	finishedAt := time.Date(2026, 6, 8, 14, 10, 0, 0, time.UTC)
-
-	runningEvents := BuildCanonicalSessionEvents(
-		SessionReadResult{
-			SessionID:        "dur-sess-js-run-n-001",
-			Status:           LifecycleStatusRunning,
-			OrchestratorKind: "JAVASCRIPT",
-			Dialect:          "you-workflow-v1",
-			Phase:            "verify",
-			Lifecycle:        &LifecycleTimestamps{StartedAt: &startedAt},
-		},
-		ResultReadResult{
-			SessionID:    "dur-sess-js-run-n-001",
-			ResultStatus: ResultStatusPartial,
-		},
-	)
-	if len(runningEvents) != 2 {
-		t.Fatalf("running events = %d, want 2", len(runningEvents))
-	}
-	assertCanonicalEventEnvelope(t, runningEvents[0], "SESSION_STARTED", "session-started/dur-sess-js-run-n-001")
-	assertCanonicalEventEnvelope(t, runningEvents[1], "SESSION_RESULT_UPDATED", "session-result-updated/dur-sess-js-run-n-001")
-
-	terminalEvents := BuildCanonicalSessionEvents(
-		SessionReadResult{
-			SessionID:        "dur-sess-js-success-002",
-			Status:           LifecycleStatusSucceeded,
-			OrchestratorKind: "JAVASCRIPT",
-			Lifecycle:        &LifecycleTimestamps{StartedAt: &startedAt, FinishedAt: &finishedAt},
-		},
-		ResultReadResult{
-			SessionID:    "dur-sess-js-success-002",
-			ResultStatus: ResultStatusFinal,
-		},
-	)
-	if len(terminalEvents) != 3 {
-		t.Fatalf("terminal events = %d, want 3", len(terminalEvents))
-	}
-	assertCanonicalEventEnvelope(t, terminalEvents[2], "SESSION_COMPLETED", "session-completed/dur-sess-js-success-002")
-}
-
-func TestFilterEventsAfterReconnect_AfterEventIDAndSequence(t *testing.T) {
-	events := []json.RawMessage{
-		json.RawMessage(`{"id":"session-started/s1","context":{"sequence":1,"sessionSequence":0}}`),
-		json.RawMessage(`{"id":"session-result-updated/s1","context":{"sequence":2,"sessionSequence":1}}`),
-		json.RawMessage(`{"id":"session-completed/s1","context":{"sequence":3,"sessionSequence":2}}`),
-	}
-
-	all, err := FilterEventsAfterReconnect(events, EventReconnectRequest{}, "s1")
-	if err != nil {
-		t.Fatalf("FilterEventsAfterReconnect all: %v", err)
-	}
-	if len(all) != 3 {
-		t.Fatalf("all events = %d, want 3", len(all))
-	}
-
-	afterStart, err := FilterEventsAfterReconnect(events, EventReconnectRequest{
-		AfterEventID: "session-started/s1",
-	}, "s1")
-	if err != nil {
-		t.Fatalf("FilterEventsAfterReconnect after start: %v", err)
-	}
-	if len(afterStart) != 2 {
-		t.Fatalf("after start events = %d, want 2", len(afterStart))
-	}
-
-	sequence := 1
-	afterSequence, err := FilterEventsAfterReconnect(events, EventReconnectRequest{
-		AfterSequence: &sequence,
-	}, "s1")
-	if err != nil {
-		t.Fatalf("FilterEventsAfterReconnect after sequence: %v", err)
-	}
-	if len(afterSequence) != 1 {
-		t.Fatalf("after sequence events = %d, want 1", len(afterSequence))
-	}
-
-	_, err = FilterEventsAfterReconnect(events, EventReconnectRequest{
-		AfterEventID: "missing-event",
-	}, "s1")
-	if !errors.Is(err, ErrReconnectCursorNotFound) {
-		t.Fatalf("missing cursor error = %v, want ErrReconnectCursorNotFound", err)
-	}
-}
-
-func assertCanonicalEventEnvelope(t *testing.T, raw json.RawMessage, eventType, id string) {
-	t.Helper()
-	var envelope struct {
-		SchemaVersion string `json:"schemaVersion"`
-		ID            string `json:"id"`
-		Type          string `json:"type"`
-		Context       struct {
-			Sequence int `json:"sequence"`
-		} `json:"context"`
-		Payload json.RawMessage `json:"payload"`
-	}
-	if err := json.Unmarshal(raw, &envelope); err != nil {
-		t.Fatalf("Unmarshal event: %v", err)
-	}
-	if envelope.SchemaVersion != canonicalFactoryEventSchemaVersion {
-		t.Fatalf("schemaVersion = %q, want %q", envelope.SchemaVersion, canonicalFactoryEventSchemaVersion)
-	}
-	if id != "" && envelope.ID != id {
-		t.Fatalf("id = %q, want %q", envelope.ID, id)
-	}
-	if eventType != "" && envelope.Type != eventType {
-		t.Fatalf("type = %q, want %q", envelope.Type, eventType)
-	}
-	if envelope.Context.Sequence <= 0 {
-		t.Fatalf("sequence = %d, want positive", envelope.Context.Sequence)
-	}
-	if len(envelope.Payload) == 0 {
-		t.Fatal("payload missing")
-	}
-}
