--- pkg/factorysessionexecution/result_projection_test.go
@@ -1,169 +1,0 @@
-package factorysessionexecution
-
-import (
-	"context"
-	"encoding/json"
-	"testing"
-)
-
-func TestProjectResultRead_ModePartialAndFinal(t *testing.T) {
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-js-run-n-001")
-
-	partial, err := service.GetResult(context.Background(), "dur-sess-js-run-n-001", ResultRequest{Mode: ResultModePartial})
-	if err != nil {
-		t.Fatalf("GetResult partial: %v", err)
-	}
-	if partial.ResultStatus != ResultStatusPartial {
-		t.Fatalf("partial status = %q, want PARTIAL", partial.ResultStatus)
-	}
-	if len(partial.PrimaryResult) == 0 {
-		t.Fatal("partial primaryResult missing")
-	}
-	if partial.Mode != ResultModePartial {
-		t.Fatalf("mode = %q, want partial", partial.Mode)
-	}
-
-	final, err := service.GetResult(context.Background(), "dur-sess-js-run-n-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult final: %v", err)
-	}
-	if final.ResultStatus != ResultStatusNotReady {
-		t.Fatalf("final status = %q, want NOT_READY", final.ResultStatus)
-	}
-	if len(final.PrimaryResult) != 0 {
-		t.Fatal("final primaryResult should be omitted for running session")
-	}
-	if final.Availability == nil || final.Availability.Reason != "RESULT_NOT_READY" {
-		t.Fatalf("availability = %#v, want RESULT_NOT_READY", final.Availability)
-	}
-}
-
-func TestProjectResultRead_TerminalFinalAndUnavailable(t *testing.T) {
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-petri-success-001")
-
-	final, err := service.GetResult(context.Background(), "dur-sess-petri-success-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult terminal final: %v", err)
-	}
-	if final.ResultStatus != ResultStatusFinal {
-		t.Fatalf("status = %q, want FINAL", final.ResultStatus)
-	}
-	if len(final.PrimaryResult) == 0 {
-		t.Fatal("final primaryResult missing")
-	}
-
-	startAsyncByRequestID(t, service, "req-petri-cancel-001")
-	unavailable, err := service.GetResult(context.Background(), "dur-sess-petri-cancel-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult unavailable: %v", err)
-	}
-	if unavailable.ResultStatus != ResultStatusUnavailable {
-		t.Fatalf("status = %q, want UNAVAILABLE", unavailable.ResultStatus)
-	}
-	if unavailable.Availability == nil || unavailable.Availability.Reason != "SESSION_CANCELED" {
-		t.Fatalf("availability = %#v", unavailable.Availability)
-	}
-}
-
-func TestProjectResultRead_FailedWithPartialHonorsPartialMode(t *testing.T) {
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-js-failed-partial-001")
-
-	result, err := service.GetResult(context.Background(), "dur-sess-js-failed-partial-001", ResultRequest{Mode: ResultModePartial})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.ResultStatus != ResultStatusFailedWithPartial {
-		t.Fatalf("status = %q, want FAILED_WITH_PARTIAL", result.ResultStatus)
-	}
-	if len(result.PrimaryResult) == 0 {
-		t.Fatal("partial primaryResult missing")
-	}
-	if result.Failure == nil || !result.Failure.PartialResultAvailable {
-		t.Fatal("failure detail missing")
-	}
-}
-
-func TestProjectResultRead_IncludeArtifactsShaping(t *testing.T) {
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-petri-success-001")
-
-	excluded, err := service.GetResult(context.Background(), "dur-sess-petri-success-001", ResultRequest{
-		Mode:             ResultModeFinal,
-		IncludeArtifacts: false,
-	})
-	if err != nil {
-		t.Fatalf("GetResult excluded: %v", err)
-	}
-	if excluded.IncludeArtifacts {
-		t.Fatal("includeArtifacts = true, want false")
-	}
-	if len(excluded.ArtifactRefs) != 0 {
-		t.Fatalf("artifactRefs = %#v, want omitted", excluded.ArtifactRefs)
-	}
-	if len(excluded.ArtifactIDs) != 1 || excluded.ArtifactIDs[0] != "art-petri-final-001" {
-		t.Fatalf("artifactIds = %#v", excluded.ArtifactIDs)
-	}
-
-	included, err := service.GetResult(context.Background(), "dur-sess-petri-success-001", ResultRequest{
-		Mode:             ResultModeFinal,
-		IncludeArtifacts: true,
-	})
-	if err != nil {
-		t.Fatalf("GetResult included: %v", err)
-	}
-	if !included.IncludeArtifacts {
-		t.Fatal("includeArtifacts = false, want true")
-	}
-	if len(included.ArtifactRefs) != 1 || included.ArtifactRefs[0].ID != "art-petri-final-001" {
-		t.Fatalf("artifactRefs = %#v", included.ArtifactRefs)
-	}
-	if len(included.ArtifactIDs) != 0 {
-		t.Fatalf("artifactIds = %#v, want omitted when refs included", included.ArtifactIDs)
-	}
-}
-
-func TestProjectResultRead_NotReadyRunningSession(t *testing.T) {
-	service := newContractFakeService(t)
-	startAsyncByRequestID(t, service, "req-petri-run-001")
-
-	result, err := service.GetResult(context.Background(), "dur-sess-petri-run-001", ResultRequest{Mode: ResultModeFinal})
-	if err != nil {
-		t.Fatalf("GetResult: %v", err)
-	}
-	if result.ResultStatus != ResultStatusNotReady {
-		t.Fatalf("status = %q, want NOT_READY", result.ResultStatus)
-	}
-	if result.Availability == nil || result.Availability.Message == "" {
-		t.Fatal("availability missing")
-	}
-}
-
-func TestProjectResultRead_DefaultsToFinalMode(t *testing.T) {
-	canonical := ResultReadResult{
-		SessionID:     "dur-sess-001",
-		ResultStatus:  ResultStatusFinal,
-		SessionStatus: LifecycleStatusSucceeded,
-		PrimaryResult: json.RawMessage(`[{"type":"text","text":"done"}]`),
-	}
-	session := SessionReadResult{
-		SessionID: "dur-sess-001",
-		Status:    LifecycleStatusSucceeded,
-		ResultSummary: &ResultSummary{
-			ResultStatus: string(ResultStatusFinal),
-		},
-	}
-
-	projected, err := ProjectResultRead(canonical, session, nil, ResultRequest{})
-	if err != nil {
-		t.Fatalf("ProjectResultRead: %v", err)
-	}
-	if projected.Mode != ResultModeFinal {
-		t.Fatalf("mode = %q, want final", projected.Mode)
-	}
-	if projected.ResultStatus != ResultStatusFinal {
-		t.Fatalf("status = %q, want FINAL", projected.ResultStatus)
-	}
-}
